+++ core/modules/modScrumProject.class.php
@@ -382,6 +382,26 @@ class modScrumProject extends DolibarrModules
 		$this->rights[$r][5] = 'delete'; // In php code, permission will be checked by test if ($user->rights->scrumproject->level1->level2)
 		$r++;
 
+		/**
+		 * DROIT POUR KANBAN
+		 */
+
+		$this->rights[$r][0] = $this->numero . $r; // Permission id (must not be already used)
+		$this->rights[$r][1] = 'ReadScrumKanban'; // Permission label
+		$this->rights[$r][4] = 'scrumkanban'; // In php code, permission will be checked by test if ($user->rights->scrumproject->level1->level2)
+		$this->rights[$r][5] = 'read'; // In php code, permission will be checked by test if ($user->rights->scrumproject->level1->level2)
+		$r++;
+		$this->rights[$r][0] = $this->numero . $r; // Permission id (must not be already used)
+		$this->rights[$r][1] = 'WriteScrumKanban'; // Permission label
+		$this->rights[$r][4] = 'scrumkanban'; // In php code, permission will be checked by test if ($user->rights->scrumproject->level1->level2)
+		$this->rights[$r][5] = 'write'; // In php code, permission will be checked by test if ($user->rights->scrumproject->level1->level2)
+		$r++;
+		$this->rights[$r][0] = $this->numero . $r; // Permission id (must not be already used)
+		$this->rights[$r][1] = 'DeleteScrumKanban'; // Permission label
+		$this->rights[$r][4] = 'scrumkanban'; // In php code, permission will be checked by test if ($user->rights->scrumproject->level1->level2)
+		$this->rights[$r][5] = 'delete'; // In php code, permission will be checked by test if ($user->rights->scrumproject->level1->level2)
+		$r++;
+
 		// Main menu entries to add
 		$this->menu = array();
 		$r = 0;